April 27, 2017 Update
With this update it is now possible to create a basic character (!new
char), start a test adventure (!adventure start test), have others join
it (!enter), and do various adventure action commands such as !go,
!look, !pull, !read.

There's also a very simplistic battle combat system added for testing.

Still a lot of work left.. but it's a start.