April 28, 2017 update
!new char <race> can now have human, elf, or galka

Basic stats added. Starting Stats are based on the race you pick and
will be used for every level 1 job.

Lots of work done on viewing current and resting stats.

Started copying the FFXIV damage formula over to this game. It might not
be 100% but it's close enough for what I need it for.

Battle system improved slightly.

Still a lot to do but it's chugging along.