April 30, 2017 update
Added code to check for level up

Added code to check for level 60 level cap and the soft cap of 50. It
checks to see if the Genkai adventure is clear.

Added code for increasing the number of battles/adventures and
wins/losses of each.

Fixed adventures to write the $fulldate when the very first adventure is
started