May 1, 2017 update
Added the rest of the !rest command. It now works properly.

Added the ability for adventures to properly give XP to players

Added the ability for adventures to properly give fame to players for
clearing the adventure

Added the ability for adventures to properly mark that the player has
cleared the adventure if it's a victory

Added the ability for players to level up properly

Added the ability for everyone to regain 10% TP per turn (currently 100)

Fixed a small issue with the melee damage formula not counting the stat
properly