May 1, 2017 Update - Part 2
Changed !look a little bit. Outside of an adventure it shows what you're
wearing to people. Inside of adventure but not in battle it shows the
room. Inside of battle it shows the battle info.

Changed the char file to store the main currency as "money" rather than
"gil" in case other bot owners don't want to use gil.

Changed where the xp/money/item drops are stored in the monster files