Added !spells and !abilities
Use these commands to see what spells and abilities are available to
you.  If you have colors on you can tell if you have enough mp/tp for it
or if the spell/abiliity is still on cooldown.  Bright red means not
enough tp/mp and maroon means it's on cooldown.